UserService/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==7.4.2
pytest-flask==1.2.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
